requirements.txt
Pillow==10.4.0
requests==2.28.1
